--- a/src/lu.jl
+++ b/src/lu.jl
@@ -144,7 +144,6 @@ end
 end
 
 @inline function nsplit(::Type{T}, n) where {T}
-    # k = max(2, 512 ÷ (isbitstype(T) ? sizeof(T) : 8))
     k = max(2, 128 ÷ (isbitstype(T) ? sizeof(T) : 8))
     k_2 = k ÷ 2
     return n >= k ? ((n + k_2) ÷ k) * k_2 : n ÷ 2