After:
```julia
julia> @benchmark RecursiveFactorization.lu!(copyto!($B,$A), $ipiv,Val(true),Val(false), blocksize=8)
BenchmarkTools.Trial: 2562 samples with 1 evaluation.
 Range (min … max):  384.506 μs … 399.233 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     386.841 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   387.488 μs ±   1.524 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

                ▃▂▄▄▅█▅▂▁▂                                       
  ▂▁▁▁▂▂▂▃▃▄▆▅▇████████████▇▅▄▄▃▃▃▃▃▃▃▄▄▅▄▆▆▇▇▇▇▇▇█▆▇▆▇▅▅▄▃▃▃▃▃ ▄
  385 μs           Histogram: frequency by time          391 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
Before:
```julia
julia> @benchmark RecursiveFactorization.lu!(copyto!($B,$A), $ipiv,Val(true),Val(false), blocksize=8)
BenchmarkTools.Trial: 2318 samples with 1 evaluation.
 Range (min … max):  426.292 μs … 445.566 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     428.158 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   428.777 μs ±   1.491 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

           ▁▂▄▆▆█▅▆▆▆▁▃▁                    ▁▁▁▁▃▁ ▂▁            
  ▂▂▂▂▃▅▄▅██████████████▇▇▄▃▃▂▁▂▂▂▂▂▃▅▄▄▄▅▇▆██████▆██▇▆▆▄▄▃▃▃▃▃ ▄
  426 μs           Histogram: frequency by time          432 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
Setup:
```julia
@time using RecursiveFactorization
A = rand(300,300);
B = similar(A);
ipiv = Vector{Int}(undef,300);
```

Benchmarks across all sizes `4:500`, RF 0.2.19:
![lubenchrf19](https://github.com/JuliaLinearAlgebra/RecursiveFactorization.jl/assets/8043603/fee9812f-d610-4e91-bb3b-6a11bad94e23)
![lubenchturbofindma](https://github.com/JuliaLinearAlgebra/RecursiveFactorization.jl/assets/8043603/82468c92-0b72-41d6-8f2d-a15f98770e99)
versus PR on bottom.

Summary stats of relative differences:
```julia
julia> StatsBase.summarystats(respr ./ res19)
Summary Stats:
Length:         497
Missing Count:  0
Mean:           1.086197
Minimum:        0.889275
1st Quartile:   1.059688
Median:         1.092518
3rd Quartile:   1.120472
Maximum:        1.254754
```
So, 8.6% increase in GFLOPS on average on
```julia
julia> versioninfo()
Julia Version 1.9.3-DEV.0
Commit 6fc1be04ee (2023-07-06 14:55 UTC)
Platform Info:
  OS: Linux (x86_64-generic-linux)
  CPU: 36 × Intel(R) Core(TM) i9-10980XE CPU @ 3.00GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, cascadelake)
  Threads: 1 on 36 virtual cores
```