@chriselrod 2023-08-09
> This looks pretty good, but the test failure looks real.

It is. It's specific to x86_64 CPUs without AVX2.