@chriselrod 2023-08-09
For a long time, I tried to get good performance on them when using integers by using `Int32` instead of `Int64`, but that introduced bugs. I think I should just not care about performance and use `Int64`.